requirements.txt
flask
tensorflow
keras
numpy
pillow
opencv-python
matplotlib
scikit-learn
efficientnet